have to fix bgs

+++ src/com/company/Algorithems.java
@@ -1,5 +1,5 @@
 package com.company;
 
 public interface Algorithems {
-    int solve(int[] array);
+    float solve(Float[] array);
 }
